Don't try to rename it if it doesn't exist
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@16057 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/support/FileLoggerHook.java
@@ -503,7 +503,8 @@ public class FileLoggerHook extends LoggerHook {
 		java.util.Arrays.sort(files);
 		long lastStartTime = -1;
 		File oldFile = null;
-        FileUtil.renameTo(latestFile, previousFile);
+        if(latestFile.exists())
+        	FileUtil.renameTo(latestFile, previousFile);
 		boolean logMINOR = Logger.shouldLog(Logger.MINOR, this);
 		for(int i=0;i<files.length;i++) {
 			File f = files[i];
